+++ apps/mobile/src/store/subscription/store.ts
@@ -3,7 +3,7 @@ import { FeedViewType } from "@follow/constants"
 import type { SubscriptionSchema } from "@/src/database/schemas/types"
 import { apiClient } from "@/src/lib/api-fetch"
 import { honoMorph } from "@/src/morph/hono"
-import { storeDbMorph } from "@/src/morph/store-db"
+import { buildSubscriptionDbId, storeDbMorph } from "@/src/morph/store-db"
 import { SubscriptionService } from "@/src/services/subscription"
 
 import { feedActions } from "../feed/store"
@@ -115,19 +115,35 @@ class SubscriptionActions {
       draft.subscriptionIdSet = new Set()
     })
   }
+
+  reset() {
+    immerSet((draft) => {
+      Object.assign(draft, defaultState)
+    })
+  }
 }
 
 class SubscriptionSyncService {
-  async fetch(view: FeedViewType) {
+  async fetch(view?: FeedViewType) {
     const res = await apiClient.subscriptions.$get({
       query: {
         view: view !== undefined ? String(view) : undefined,
       },
     })
 
-    subscriptionActions.resetByView(view)
+    if (typeof view === "number") {
+      subscriptionActions.resetByView(view)
+    } else {
+      subscriptionActions.reset()
+    }
 
     const { subscriptions, feeds, lists, inboxes } = honoMorph.toSubscription(res.data)
+
+    await SubscriptionService.deleteNotExists(
+      subscriptions.map((s) => buildSubscriptionDbId(s)),
+      view,
+    )
+
     feedActions.upsertMany(feeds)
     subscriptionActions.upsertMany(subscriptions)
     listActions.upsertMany(lists)
@@ -250,7 +266,7 @@ class SubscriptionSyncService {
     })
 
     tx.persist(() => {
-      return SubscriptionService.delete(storeDbMorph.toSubscriptionSchema(subscription))
+      return SubscriptionService.delete(buildSubscriptionDbId(subscription))
     })
 
     tx.request(async () => {