+++ apps/renderer/src/hooks/biz/useTimelineList.ts
@@ -1,20 +1,21 @@
 import { useMemo } from "react"
 
-import { useAllInboxes, useAllLists } from "~/store/subscription/hooks"
+import { useAllInboxes, useAllLists, useViewWithSubscription } from "~/store/subscription/hooks"
 
-const views = [0, 1, 2, 3].map((view) => `view-${view}`)
 export const useTimelineList = () => {
   const lists = useAllLists()
   const inboxes = useAllInboxes()
+  const views = useViewWithSubscription()
 
   const listsIds = useMemo(() => lists.map((list) => `list-${list.listId}`), [lists])
   const inboxesIds = useMemo(() => inboxes.map((inbox) => `inbox-${inbox.inboxId}`), [inboxes])
+  const viewsIds = useMemo(() => views.map((view) => `view-${view}`), [views])
 
   return {
-    views,
+    views: viewsIds,
     lists: listsIds,
     inboxes: inboxesIds,
 
-    all: useMemo(() => [...views, ...inboxesIds, ...listsIds], [listsIds, inboxesIds]),
+    all: useMemo(() => [...viewsIds, ...inboxesIds, ...listsIds], [viewsIds, listsIds, inboxesIds]),
   }
 }